Reset existing overrides when adjusting cruise speed.
If the setpoint is adjusted directly (with a script argument or a hotkey), then the script can go from accelerating to decelerating or vice versa without passing through the deadzone and without triggering the branches that handle forwards/backwards inputs. When this happens, the script doesn't reset existing overrides, but does start controlling the opposing thrusters from its previous acceleration.

This change causes it to clear those overrides (blindly), so that cruise always seeks towards the setpoint without fighting its own overrides. No similar changes are needed for the governor and decoupled modes, as they do not respect commanded speeds: all speed changes in those modes require the user to hit a control direction, which triggers the logic in input handling that clears overrives.